Put source code and tests in separate directories
