+++ src/main.cpp
@@ -1,0 +1,11 @@
+#include <iostream>
+#include "server.hpp"
+
+int main()
+{
+    LogImpl logImpl(std::cin, "logs.txt");
+
+    logImpl.process();
+
+    return 0;
+}
