+++ test/CMakeLists.txt
@@ -1,6 +1,27 @@
 cmake_minimum_required(VERSION 3.8)
 
+# --------------------------------------------------
+# Set up GoogleTest
+# --------------------------------------------------
+
+# Include GoogleTest
+include(FetchContent)
+FetchContent_Declare(
+  googletest
+  GIT_REPOSITORY https://github.com/google/googletest.git
+  GIT_TAG release-1.12.1
+)
+
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+FetchContent_MakeAvailable(googletest)
+
+enable_testing()
+
+# --------------------------------------------------
+
 set(This serverTests)
+include_directories(../src)
 
 set(Sources
     serverTests.cpp
@@ -13,8 +34,11 @@ target_link_libraries(${This} PUBLIC
     server
 )
 
-add_test(
-    NAME ${This}
-    COMMAND ${This}
-)
+# add_test(
+#     NAME ${This}
+#     COMMAND ${This}
+# )
 
+# Enable CMake's test runner to discover the tests included in the binary
+include(GoogleTest)
+gtest_discover_tests(${This})
